% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_create}
\alias{db_create}
\title{Create A Database Handle}
\usage{
db_create(dbenv = NULL, flags = 0L)
}
\arguments{
\item{dbenv}{A DB_ENV handle created with db_env_create, or NULL}

\item{flags}{flags parameter created with mkFlags}
}
\value{
A database handle.
}
\description{
Create a database handle
}
\details{
See Berkeley DB API documentation for usage details.
}
\examples{

\dontrun{
dbh <- db_create()
}

}
\author{
Jeffrey A. Ryan
}
\references{
Oracle Berkeley DB API\cr
\url{http://www.oracle.com/technetwork/database/database-technologies/berkeleydb/overview/index.html}\cr\cr
}
\seealso{
\code{\link{Databases}}\cr \code{\link{db_env_create}}
}
\keyword{database}

