\name{RBerkeley-package}
\Rdversion{1.1}
\alias{RBerkeley-package}
\docType{package}
\title{
R Interface to Embedded Oracle Berkeley DB
}
\description{
Package provides a low and high level interface
to the C API of the Oracle Berkeley DB embedded
database product.
}
\details{
Initial implementation will be to provide direct
access to the low-level API calls in their
entirety.

The project will eventually include a higher-level
interface to make using the DB from R easier.
}
\author{
Jeffrey A. Ryan
}
\references{
Oracle Berkey DB 4.7.25 \url{http://www.oracle.com/technology/documentation/berkeley-db/db/index.html}
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
