% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/env.R
\name{db_env_create}
\alias{db_env_create}
\title{Create An Environment Handle}
\usage{
db_env_create(flags = 0L)
}
\arguments{
\item{flags}{flags parameter created with mkFlags}
}
\value{
A database environment handle.
}
\description{
Create An Environment Handle
}
\details{
See Berkeley DB API documentation for usage details.
}
\examples{
\dontrun{
dbh <- db_env_create()
}
}
\author{
Jeffrey A. Ryan
}
\references{
Oracle Berkeley DB API\cr
\url{http://www.oracle.com/technetwork/database/database-technologies/berkeleydb/overview/index.html}\cr
}
\seealso{
\code{\link{Environments}}\cr \code{\link{db_create}}
}
\keyword{database}

